+++ src/main/java/io/delivery/config/SecurityConfig.java
@@ -34,8 +34,8 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter{
         http.authorizeRequests()
                 .antMatchers("/").permitAll()
                 .antMatchers("/secure").access("hasRole('ADMIN')")
-                .antMatchers("/backup").access("hasRole('ADMIN')")
                 .antMatchers("/dump").access("hasRole('ADMIN')")
+//                .antMatchers("/admin/*").access("hasRole('ADMIN')")
                 .and().csrf().disable().formLogin().defaultSuccessUrl("/", false);
     }
